+++ .gitignore
@@ -1,2 +1,4 @@
 
 __pycache__/player.cpython-310.pyc
+__pycache__/player.cpython-310.pyc
+*.pyc
