+++ README.md
@@ -1,2 +1,2 @@
 # CSE-350-Connect-4-Project
-Project repository for CSE 350 Term Project on creating a Connect-4 Game
+Branch for me to try some things and commit anything I feel like without having to worry
